CLIENT/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.71.0",
    "PySide6>=6.6.2",
    "protobuf==5.29.0",
]
